1、添加是否删除整轨和是否删除转码前的音频的设置 2、增加了一点日志输出 3、更新了readme.md

+++ lib/audio_handler.py
@@ -1,4 +1,4 @@
-import subprocess, json, os, io,re
+import subprocess, json, os, io,re, logging
 import wave
 import mutagen
 from mutagen.flac import Picture
@@ -7,32 +7,46 @@ from PIL import Image
 from lib.metadata_mapping import ID3v24_TO_VORBIS, APEV2_TO_VORBIS, MP4_TO_VORBIS
 from lib.common_method import *
 
+logger = logging.getLogger(__name__)
 
 os.environ['PATH'] = os.environ['PATH'] + os.pathsep + os.getcwd() + '/encoders/'
 
 ALLOWED_SAMPLE_FMT = ['s32', 's32p', 's16', 's16p']
 
+EXCLUDED_DIRS = ['bk', 'booklet', 'scan', 'scans', 'artwork', 'artworks']
 
 class AudioHandler:
     @staticmethod
-    def convert_wav_to_flac(file_path, root, name):
+    def common_cmd2flac(cmd, file_path, root, name):
+        subprocess.run(cmd, capture_output=True, text=True, check=True)
+        logger.info(f'成功将{file_path}转换为wav')
+        logger.info(f'正在将{os.path.join(root, name)}.wav转换为flac')
+        AudioHandler.convert_wav_to_flac(f'{os.path.join(root, name)}.wav', root, name)
+        logger.info(f'成功将{os.path.join(root, name)}.wav转换为flac')
+        os.remove(f'{os.path.join(root, name)}.wav')
+        logger.info(f'删除缓存的{os.path.join(root, name)}.wav')
+
+    @staticmethod
+    def convert_wav_to_flac(file_path, root, name):  # 缓存文件转换为flac
         try:
             cmd = ['flac.exe', file_path, '--best', '--threads=16', '-o', f'{os.path.join(root, name)}.flac']
             subprocess.run(cmd, capture_output=True, text=True, check=True)
         except subprocess.CalledProcessError as e:
-            print(e.stderr)
-            raise Exception('未知错误')
+            logger.error(e.stdout)
+            raise Exception('还没搞异常处理，暂时打印出来')
 
     @staticmethod
     def tmp_file_save_to_wav(root, track, channels, bits_per_sample, sample_rate, raw):
         filename = f"{track['TRACKNUMBER']} - {track['TITLE']}.wav"
         filename = re.sub(r'[<>:"/\\|?*]', '_', filename)
         file_path = os.path.join(root, filename)
+        logger.debug(f'将分轨文件暂存到{file_path}')
         with wave.open(file_path, 'wb') as wf:
             wf.setnchannels(channels)  # 设置为立体声（两个通道）
             wf.setsampwidth(bits_per_sample // 8)  # 每个采样点2字节（16位）
             wf.setframerate(sample_rate)  # 设置采样率
             wf.writeframes(raw)
+        logger.debug(f'分轨文件暂存完成')
         return file_path
 
     @staticmethod
@@ -42,67 +56,78 @@ class AudioHandler:
             result = subprocess.run(cmd, capture_output=True, text=True, check=True)
         except subprocess.CalledProcessError as e:
             print(e.stderr)
-            raise Exception('未知错误')
+            raise Exception('还没搞异常处理，暂时打印出来')
         data = json.loads(result.stdout)
         sample_fmt = data['streams'][0]['sample_fmt']
         codec_name = data['streams'][0]['codec_name']
-        # TODO: 把所有的有损编码格式都写进去
         if codec_name == 'aac':
-            print(f'有损音频不会转换，文件{file_path}')
+            logger.info(f'有损音频不会转换，文件{file_path}')
             return False
         if sample_fmt in ALLOWED_SAMPLE_FMT:
             return True
         else:
-            print(f'不支持的编码格式，编码格式为{sample_fmt}')
+            logger.info(f'不支持的编码格式，编码格式为{sample_fmt}')
             return False
 
     @staticmethod
-    def wav2flac(file_path):
+    def wav2flac(file_path, is_delete_origin_audio):  # 源文件是wav的转换为flac
         root, name = get_file_name_and_root(file_path)
+        logger.info(f'正在将{file_path}转换为flac')
         AudioHandler.convert_wav_to_flac(file_path, root, name)
+        logger.info(f'成功将{file_path}转换为flac')
         MetaHandler.id3_tag_to_vorbis(file_path, root, name)
-        os.remove(file_path)
+        logger.info(f'成功将{file_path}的元数据转移到{os.path.join(root, name)}.flac')
+        if is_delete_origin_audio:
+            os.remove(file_path)
+            logger.info(f'删除源文件{file_path}')
 
     @staticmethod
-    def m4a2flac(file_path):
-        print(file_path)
+    def m4a2flac(file_path, is_delete_origin_audio):
+        logger.info(f'正在将{file_path}暂时转换为wav，为后续处理')
         root, name = get_file_name_and_root(file_path)
         cmd = ['refalac64.exe', '-D', file_path, '-o', f'{os.path.join(root, name)}.wav']
-        subprocess.run(cmd, capture_output=True, text=True, check=True)
-        AudioHandler.convert_wav_to_flac(f'{os.path.join(root, name)}.wav', root, name)
-        os.remove(f'{os.path.join(root, name)}.wav')
+        AudioHandler.common_cmd2flac(cmd, file_path, root, name)
         MetaHandler.mp4_to_vorbis(file_path, root, name)
-        os.remove(file_path)
+        logger.info(f'成功将{file_path}的元数据转移到{os.path.join(root, name)}.flac')
+        if is_delete_origin_audio:
+            os.remove(file_path)
+            logger.info(f'删除源文件{file_path}')
 
     @staticmethod
-    def ape2flac(file_path):
+    def ape2flac(file_path, is_delete_origin_audio):
+        logger.info(f'正在将{file_path}暂时转换为wav，为后续处理')
         root, name = get_file_name_and_root(file_path)
         cmd = ['mac.exe', file_path, f'{os.path.join(root, name)}.wav', '-d', '-threads=16']
-        subprocess.run(cmd, capture_output=True, text=True, check=True)
-        AudioHandler.convert_wav_to_flac(f'{os.path.join(root, name)}.wav', root, name)
-        os.remove(f'{os.path.join(root, name)}.wav')
+        AudioHandler.common_cmd2flac(cmd, file_path, root, name)
         MetaHandler.apev2_to_vorbis(file_path, root, name)
-        os.remove(file_path)
+        logger.info(f'成功将{file_path}的元数据转移到{os.path.join(root, name)}.flac')
+        if is_delete_origin_audio:
+            os.remove(file_path)
+            logger.info(f'删除源文件{file_path}')
 
     @staticmethod
-    def tak2flac(file_path):
+    def tak2flac(file_path, is_delete_origin_audio):
+        logger.info(f'正在将{file_path}暂时转换为wav，为后续处理')
         root, name = get_file_name_and_root(file_path)
         cmd = ['Takc.exe', '-d', file_path, f'{os.path.join(root, name)}.wav']
-        subprocess.run(cmd, capture_output=True, text=True, check=True)
-        AudioHandler.convert_wav_to_flac(f'{os.path.join(root, name)}.wav', root, name)
-        os.remove(f'{os.path.join(root, name)}.wav')
+        AudioHandler.common_cmd2flac(cmd, file_path, root, name)
         MetaHandler.apev2_to_vorbis(file_path, root, name)
-        os.remove(file_path)
+        logger.info(f'成功将{file_path}的元数据转移到{os.path.join(root, name)}.flac')
+        if is_delete_origin_audio:
+            os.remove(file_path)
+            logger.info(f'删除源文件{file_path}')
 
     @staticmethod
-    def tta2flac(file_path):
+    def tta2flac(file_path, is_delete_origin_audio):
+        logger.info(f'正在将{file_path}暂时转换为wav，为后续处理')
         root, name = get_file_name_and_root(file_path)
         cmd = ['ttaenc.exe', '-d', file_path, f'{os.path.join(root, name)}.wav']
-        subprocess.run(cmd, capture_output=True, text=True, check=True)
-        AudioHandler.convert_wav_to_flac(f'{os.path.join(root, name)}.wav', root, name)
-        os.remove(f'{os.path.join(root, name)}.wav')
+        AudioHandler.common_cmd2flac(cmd, file_path, root, name)
         MetaHandler.id3_tag_to_vorbis(file_path, root, name)
-        os.remove(file_path)
+        logger.info(f'成功将{file_path}的元数据转移到{os.path.join(root, name)}.flac')
+        if is_delete_origin_audio:
+            os.remove(file_path)
+            logger.info(f'删除源文件{file_path}')
 
 
 HANDLERS = {
@@ -283,39 +308,52 @@ class Splitter:
         return sample_rate, channels, bits_per_sample
 
     @staticmethod
-    def split_flac_with_cue(file_path):
+    def split_flac_with_cue(file_path, is_delete_single_track):
         source_root, source_name = get_file_name_and_root(file_path)
         if os.path.exists(f'{os.path.join(source_root, source_name)}.cue'):
-
             tracks = CueParser.paser_cue_data(f'{os.path.join(source_root, source_name)}.cue')
-
             sample_rate, channels, bits_per_sample = Splitter.get_audio_data(file_path)
-
+            logger.info(f"正在将{file_path}转换为pcm数据缓存到内存中")
             process_decode = subprocess.Popen(['flac.exe', '-d', '--stdout', file_path], stdout=subprocess.PIPE,
                                        stderr=subprocess.PIPE)
-            # 读取标准输出中的 PCM 数据
             pcm_data, _ = process_decode.communicate()
-
+            logger.info(f"成功转换{file_path}为pcm数据")
             for i, track in enumerate(tracks):
+                try:
+                    logger.info(f'正在分割第{i+1}首曲目，曲目名为{track["TITLE"]}')
+                except KeyError:
+                    logger.info(f'正在分割第{i + 1}首曲目，该音频无曲目名')
+
                 start_frame = track['INDEX01']
                 end_frame = tracks[i + 1]['INDEX01'] if i < len(tracks) - 1 else None
                 raw = Splitter.extract_pcm_segment_frame(pcm_data, sample_rate, bits_per_sample, channels, start_frame,
                                                          end_frame)
 
+                logger.info('正在将曲目转换成wav缓存到硬盘')
                 split_audio_path_wav = AudioHandler.tmp_file_save_to_wav(source_root, track, channels, bits_per_sample,
                                                                      sample_rate, raw)
+                logger.info('缓存成功')
                 split_audio_root, split_audio_name = get_file_name_and_root(split_audio_path_wav)
-
+                logger.info('正在将曲目转换成flac')
                 AudioHandler.convert_wav_to_flac(split_audio_path_wav, split_audio_root, split_audio_name)
+                logger.info('成功转换为flac')
                 split_audio_path_flac = f"{os.path.join(split_audio_root, split_audio_name)}.flac"
                 split_audio_flac = mutagen.File(split_audio_path_flac)
                 for field, tag in track.items():
                     if field != 'INDEX01':
                         split_audio_flac[field] = str(tag)
                 split_audio_flac.save()
+                logger.info('成功将cuesheet的元数据写入分轨音频')
                 os.remove(split_audio_path_wav)
+                logger.info('成功删除缓存的wav')
+            if is_delete_single_track:
+                os.remove(f'{os.path.join(source_root, source_name)}.cue')
+                os.remove(file_path)
+                logger.info('成功删除源文件和源文件的cue')
+            logger.info(f'{file_path}分轨完成')
+            print('-'*50)
         else:
-            print(f'{file_path}没有cue文件')
+            logger.debug(f'{file_path}没有cue文件')
 
 
 class CueParser:
@@ -391,7 +429,6 @@ class CueParser:
                         current_track[field] = tag
 
         tracks_info.append(current_track)
-        # album_info["TOTALTRACKS"] = str(len(tracks_info))
-        album_info["TRACKTOTAL"] = str(len(tracks_info))
+        album_info["TOTALTRACKS"] = str(len(tracks_info))
         tracks_info = [i | album_info for i in tracks_info]
         return tracks_info
